Update python-package.yml
After running the original site choice of the Polygon article through multiple attempts in command prompt, Jupiter, and vs code to try to initiate soup, and failing for hours, switched to myanimelist. The new site was chosen for more immediate information on the page and is running, but smaller elements are harder to break down and sort. May need to consider another alternative given the base-level functioning code. Also, compare results for Schedule Builder reference to the scale and size of the personal desired end result.