Requests VS Code Error
Had code running in VS at one point but then had errors when trying again today. Was able to resolve one by clearing the cache, but the other only ran when I changed to 3.19.9. This ran but had less satisfying results than when it ran before. Need to find a way to ensure it runs consistently to move on to organizing pulled information.